extract export utilities
